Rename `lib/model` -> `lib/models` and `networking/apis` -> `networking/api`
