+++ lib/models/course/course_model.dart
