+++ lib/models/error_model.dart
